chore(splash-feature): move splash to its own file
